add support for purge_levels and purge_whitelist

+++ puppetfile.go
@@ -118,7 +118,7 @@ func resolvePuppetEnvironment(envBranch string, tags bool, outputNameTag string)
 							targetDir = normalizeDir(targetDir)
 
 							env := strings.Replace(strings.Replace(targetDir, sa.Basedir, "", 1), "/", "", -1)
-							syncToModuleDir(workDir, targetDir, branch, false, false, env)
+							syncToModuleDir(workDir, targetDir, branch, false, false, env, true)
 							if !fileExists(targetDir + "Puppetfile") {
 								Debugf("Skipping branch " + source + "_" + branch + " because " + targetDir + "Puppetfile does not exist")
 							} else {
@@ -126,6 +126,9 @@ func resolvePuppetEnvironment(envBranch string, tags bool, outputNameTag string)
 								puppetfile.workDir = normalizeDir(targetDir)
 								puppetfile.controlRepoBranch = branch
 								mutex.Lock()
+								for _, moduleDir := range puppetfile.moduleDirs {
+									desiredContent = append(desiredContent, filepath.Join(puppetfile.workDir, moduleDir))
+								}
 								allPuppetfiles[env] = puppetfile
 								allEnvironments[env] = true
 								allBasedirs[sa.Basedir] = true
@@ -152,17 +155,18 @@ func resolvePuppetEnvironment(envBranch string, tags bool, outputNameTag string)
 	//fmt.Println("allPuppetfiles: ", allPuppetfiles, len(allPuppetfiles))
 	//fmt.Println("allPuppetfiles[0]: ", allPuppetfiles["postinstall"])
 	resolvePuppetfile(allPuppetfiles)
-	//// sync to basedir
-	//for _, branch := range branches {
-	//	if len(branch) != 0 {
-	//		Debugf("Syncing branch: " + branch)
-	//		// TODO if sa.Prefix != true
-	//		if !strings.Contains(branch, "hiera") && !strings.Contains(branch, "files") {
-	//			//puppetfile := readPuppetfile(targetDir)
+	//fmt.Println(desiredContent)
+	purgeUnmanagedContent(envBranch, allBasedirs, allEnvironments)
 
-	//		}
-	//	}
-	//}
+}
+
+func purgeUnmanagedContent(envBranch string, allBasedirs map[string]bool, allEnvironments map[string]bool) {
+	if !stringSliceContains(config.PurgeLevels, "deployment") {
+		if !stringSliceContains(config.PurgeLevels, "environment") {
+			// nothing allowed to purge
+			return
+		}
+	}
 
 	for source, sa := range config.Sources {
 		prefix := resolveSourcePrefix(source, sa)
@@ -173,26 +177,63 @@ func resolvePuppetEnvironment(envBranch string, tags bool, outputNameTag string)
 
 				Debugf("Glob'ing with path " + globPath)
 				environments, _ := filepath.Glob(globPath)
-				//		environments, _ = ioutil.ReadDir(basedir)
-				//	}
-				//fmt.Println(environments)
-				//fmt.Println(allEnvironments)
+
 				for _, env := range environments {
-					envPath := strings.Split(env, "/")
-					env = envPath[len(envPath)-1]
-					Debugf("Checking if environment should exist: " + env)
-					if allEnvironments[env] {
-						Debugf("Leaving environment: " + env)
-					} else {
-						Debugf("Deleting environment: " + env)
-						if !dryRun {
-							purgeDir(basedir+env, "resolvePuppetEnvironment()")
+					if stringSliceContains(config.PurgeLevels, "environment") {
+						checkForStaleContent(env)
+					}
+					if stringSliceContains(config.PurgeLevels, "deployment") {
+						envPath := strings.Split(env, "/")
+						env = envPath[len(envPath)-1]
+						Debugf("Checking if environment should exist: " + env)
+						if allEnvironments[env] {
+							Debugf("Leaving environment: " + env)
+						} else {
+							Debugf("Deleting environment: " + env)
+							if !dryRun {
+								purgeDir(basedir+env, "purgeStaleContent()")
+							}
 						}
 					}
 				}
 			}
+		} else {
+			if stringSliceContains(config.PurgeLevels, "environment") {
+				// check for purgeable content inside -branch folder
+				checkForStaleContent(filepath.Join(sa.Basedir, prefix+envBranch))
+			}
 		}
 	}
+}
+
+func checkForStaleContent(workDir string) {
+	// add purge whitelist
+	if len(config.PurgeWhitelist) > 0 {
+		Debugf("additional purge whitelist items: " + strings.Join(config.PurgeWhitelist, " "))
+		for _, wlItem := range config.PurgeWhitelist {
+			desiredContent = append(desiredContent, filepath.Join(workDir, wlItem))
+		}
+	}
+
+	checkForStaleContent := func(path string, info os.FileInfo, err error) error {
+		stale := true
+		for _, desiredFile := range desiredContent {
+			if strings.HasPrefix(path, desiredFile) || path == workDir {
+				stale = false
+			}
+		}
+
+		if stale {
+			Infof("Removing unmanaged path " + path)
+			purgeDir(path, "checkForStaleContent()")
+		}
+		return nil
+	}
+
+	c := make(chan error)
+	Debugf("filepath.Walk'ing directory " + workDir)
+	go func() { c <- filepath.Walk(workDir, checkForStaleContent) }()
+	<-c // Walk done
 }
 
 // resolvePuppetEnvironment implements the prefix read out from each source given in the config file, like r10k https://github.com/puppetlabs/r10k/blob/master/doc/dynamic-environments/configuration.mkd#prefix
@@ -355,7 +396,7 @@ func resolvePuppetfile(allPuppetfiles map[string]Puppetfile) {
 
 				if gitModule.link {
 					Debugf("Trying to resolve " + moduleCacheDir + " with branch " + tree)
-					success = syncToModuleDir(moduleCacheDir, targetDir, tree, true, gitModule.ignoreUnreachable, env)
+					success = syncToModuleDir(moduleCacheDir, targetDir, tree, true, gitModule.ignoreUnreachable, env, false)
 				}
 
 				if len(gitModule.fallback) > 0 {
@@ -366,14 +407,14 @@ func resolvePuppetfile(allPuppetfiles map[string]Puppetfile) {
 								gitModule.ignoreUnreachable = true
 							}
 							Debugf("Trying to resolve " + moduleCacheDir + " with branch " + fallbackBranch)
-							success = syncToModuleDir(moduleCacheDir, targetDir, fallbackBranch, true, gitModule.ignoreUnreachable, env)
+							success = syncToModuleDir(moduleCacheDir, targetDir, fallbackBranch, true, gitModule.ignoreUnreachable, env, false)
 							if success {
 								break
 							}
 						}
 					}
 				} else {
-					syncToModuleDir(moduleCacheDir, targetDir, tree, gitModule.ignoreUnreachable, gitModule.ignoreUnreachable, env)
+					syncToModuleDir(moduleCacheDir, targetDir, tree, gitModule.ignoreUnreachable, gitModule.ignoreUnreachable, env, false)
 				}
 
 				// remove this module from the exisitingModuleDirs map
@@ -415,15 +456,17 @@ func resolvePuppetfile(allPuppetfiles map[string]Puppetfile) {
 	}
 	wg.Wait()
 
-	//fmt.Println(uniqueForgeModules)
-	if len(exisitingModuleDirs) > 0 && len(moduleParam) == 0 {
-		for d := range exisitingModuleDirs {
-			if strings.HasSuffix(d, ".resource_types") && isDir(d) {
-				continue
-			}
-			Debugf("Removing unmanaged file " + d)
-			if err := os.RemoveAll(d); err != nil {
-				Debugf("Error while trying to remove unmanaged file " + d)
+	if stringSliceContains(config.PurgeLevels, "puppetfile") {
+		//fmt.Println(uniqueForgeModules)
+		if len(exisitingModuleDirs) > 0 && len(moduleParam) == 0 {
+			for d := range exisitingModuleDirs {
+				if strings.HasSuffix(d, ".resource_types") && isDir(d) {
+					continue
+				}
+				Debugf("Removing unmanaged path " + d)
+				if err := os.RemoveAll(d); err != nil {
+					Debugf("Error while trying to remove unmanaged file " + d)
+				}
 			}
 		}
 	}
